Move HybridRecordEnricher in its own file
